requirements.txt
Flask
Flask-WTF
waitress
biopython
